Initial commit of files
This is a long in use project I am migrating to a new repo from a master repository that had a lot of different projects to it's own repository.

This is the FTP utility I use to push updated maps, cvar checker files, file checker configs, etc to all of the servers. ServerKeysLocal is obfuscated.

Code is very haphazard and likely needs a complete rewrite at some point. Definition of utility over good planning / implementation.